gemini_chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
pymupdf
pandas

